+++ lib/model/settings/enums/map_style.dart
@@ -17,7 +17,6 @@ extension ExtraEntryMapStyle on EntryMapStyle {
   bool get needMobileService {
     switch (this) {
       case EntryMapStyle.osmHot:
-      case EntryMapStyle.stamenToner:
       case EntryMapStyle.stamenWatercolor:
         return false;
       default: